+++ ceammc/gui/plugins/ceammc_popup_menu_ceammc.tcl
@@ -57,7 +57,7 @@
         flow.reject flow.reject_if flow.ring flow.route flow.select
         flow.seqdelay flow.space flow.speedlim flow.split flow.stack flow.sync
         flow.sync_pack flow.tee~ replace route.bang route.float route.prop
-        route.random
+        route.random route.symbol
     }}
     {flt {
         flt.a-weight flt.biquad~ flt.bpf12~ flt.bpf24~ flt.c_apf flt.c_bpf~